Refactor classes to separate files
This refactors `Processor` to `ImageProcessor` and extract it and
`Comparator` to separate files including tests.
