+++ test/eikon_test.rb
@@ -3,21 +3,8 @@ require "test_helper"
 require "byebug"
 
 class EikonTest < Minitest::Test
-  def setup
-    @image = Eikon::Processor.new("./test/images/00001.jpg")
-  end
-
   def test_that_it_has_a_version_number
     assert_not_nil ::Eikon::VERSION
-  end
-
-  def test_image_can_be_loaded
-    assert @image.nil? == false
-  end
-
-  def test_image_can_be_processed_into_byte_array
-    @image.generate_byte_array
-    assert @image.byte_array.nil? == false
   end
 
   def test_one_shot_image_loader
